requirements.txt
python-dotenv>=1.0.1
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
openai>=1.0.0
aiogram>=3.4.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
flask>=3.0.0
